submit java-web-ru/database

+++ java-web-ru/database/src/main/resources/schema.sql
@@ -4,6 +4,6 @@ DROP TABLE IF EXISTS products;
 CREATE TABLE courses (
     id INT PRIMARY KEY AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
-    price MONEY
+    price INT
 );
 -- END
